Split model auditing into their own source files
Move audit classes for BillingProjects, Accounts, MangedGroups, and
Workspaces into their own source files. Rename the remaining audit.py
file to base.py since it contains the base classes used by the classes
in the other, model-specific source files. Update import statements
and namespaces in other files.
